feat(api) : repository refactored
