+++ packages/api/src/repository/mongo/mongo-data-services.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { MongooseModule } from "@nestjs/mongoose";
-import { IDataServices } from "../../../core";
-import { DATA_BASE_CONFIGURATION } from "../../../configuration";
+import { IDataServices } from "../../core";
+import { DATA_BASE_CONFIGURATION } from "../../configuration";
 import {
   DeploymentRecord,
   DeploymentRecordSchema,
