--- packages/api/src/services/deployment-services/deployment-service.module.ts
@@ -1,8 +1,0 @@
-import { Module } from "@nestjs/common";
-import { DeploymentServicesModule } from "../../frameworks/deployment-services/deployment-service.module";
-
-@Module({
-  imports: [DeploymentServicesModule],
-  exports: [DeploymentServicesModule],
-})
-export class DeployServicesModule {}
