--- packages/api/src/services/index.ts
@@ -1,1 +1,0 @@
-// export * from './use-cases/application/application-services.use-case';
