+++ packages/api/src/use-cases/application/application-use-cases.module.ts
@@ -1,13 +1,12 @@
 import { Module } from "@nestjs/common";
 import { LoggerModule } from "src/core/logger/logger.module";
 import { DataServicesModule } from "../../services/data-services/data-services.module";
-import { DeployServicesModule } from "../../services/deployment-services/deployment-service.module";
 import { ApplicationFactoryService } from "./application-factory.service";
 import { ApplicationUseCases } from "./application.use-case";
 
 @Module({
-  imports: [DataServicesModule, DeployServicesModule],
+  imports: [DataServicesModule],
   providers: [ApplicationFactoryService, ApplicationUseCases],
   exports: [ApplicationFactoryService, ApplicationUseCases],
 })
-export class ApplicationUseCasesModule {}
+export class ApplicationUseCasesModule { }
